requirements.txt
tree-sitter
pytesseract
pillow
annotated-types==0.7.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
doc-hub @ file:///home/gimesha/My_Projects/Python/doc-hub/dist/doc_hub-1.0.0-py3-none-any.whl#sha256=34193cc10817a2d59351c1dafe5bbe7dfdaf509e270b9706d1e62a4798bec76d
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
Markdown==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pyperclip==1.11.0
pyproject_hooks==1.2.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shiboken6==6.10.0
SQLAlchemy==2.0.44
tika==3.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
Whoosh==2.7.4
